2_Data_lake_GCS_using_pandas/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
fsspec
gcsfs